Add first version of the ast
This refactors the cst a fair bit, making it take up less space, and be
easier to use.
Also fixed a bug with tokens; I was trying to be too smart for my own good.

This also adds the `InterningVec`, its a wrapper over a vec that makes
sure all items added to it are uniqued.

I'm yet to lower expressions, but it all ready to go.
I'm just done for the day.