Add stage 1 of the hir / name resolver
The hir name is from rustc, meaning High-level IR.
I'll probably rename it to name_resolver or something like that soon.

What this is:
All stage 1 does is seeding.
We effectively just state that there is something named 'x' at this
point. This is so in later stages we can just lookup a name regardless
of where or which file it was declared in.

In later stages we will actually resolve names / paths.