Finish vertical slice of `nameres::stage2`
This... was not easy.
Its a bit of a mess, but right now, I'm just happy that its in a state
to for me to start iterating on it.

Next step is to resolve locals.
Then typing,
then zir codegen.
After that we should have a nice, tall and skinny vertical slice.