Remove `pub` from most module in main
This was only done to stop rust from complaining about unused code.
Now I am trying to remove unused code.
But for the modules which are still not done I've added the
`#[allow(dead_code)]` attribute to more clearly show our intent.
Of course these do eventually need to by removed, but that's when we
actually implement the glue code to connect them all together.