zincc typer: Get some good type checking done
This is going along very nicely. So far it has been much easier then the
name resolver, I think that is because I thought this would be harder
and had put a lot more thought into it. In hindsight the name resolver
is pretty simple, but brand new problems tend to feel much harder then
they end up being.